refactor(session): move shared webui helpers out of webui package
